requirements.txt
beautifulsoup4==4.12.2
gradio==3.35.2
pick==2.2.0
python-dotenv==1.0.0
Requests==2.31.0
rich==13.4.2
